feat(objectstorage): integrate test classes related to objectstorage
