+++ server/integration-test/src/test/java/com/oceanbase/odc/service/objectstorage/LocalObjectStorageFacadeTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 OceanBase.
+ * Copyright (c) 2024 OceanBase.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -46,6 +46,7 @@ import org.springframework.web.multipart.MultipartFile;
 import com.oceanbase.odc.ServiceTestEnv;
 import com.oceanbase.odc.core.shared.exception.InternalServerError;
 import com.oceanbase.odc.service.iam.auth.AuthenticationFacade;
+import com.oceanbase.odc.service.objectstorage.LocalObjectStorageFacade;
 import com.oceanbase.odc.service.objectstorage.model.Bucket;
 import com.oceanbase.odc.service.objectstorage.model.ObjectMetadata;
 import com.oceanbase.odc.service.objectstorage.model.StorageObject;
