+++ server/integration-test/src/test/java/com/oceanbase/odc/service/objectstorage/cloud/AWSSecurityTokenServiceTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 OceanBase.
+ * Copyright (c) 2024 OceanBase.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.objectstorage.cloud;
+package com.oceanbase.odc.service.objectstorage.cloud;
 
 import java.io.File;
 
@@ -22,9 +22,12 @@ import org.junit.BeforeClass;
 import org.junit.Ignore;
 import org.junit.Test;
 
-import com.aliyun.oss.OSS;
-import com.aliyun.oss.OSSClientBuilder;
-import com.aliyun.oss.model.PutObjectResult;
+import com.amazonaws.auth.AWSSessionCredentials;
+import com.amazonaws.auth.AWSStaticCredentialsProvider;
+import com.amazonaws.auth.BasicSessionCredentials;
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.AmazonS3ClientBuilder;
+import com.amazonaws.services.s3.model.PutObjectResult;
 import com.oceanbase.odc.ITConfigurations;
 import com.oceanbase.odc.service.objectstorage.cloud.CloudObjectStorageService;
 import com.oceanbase.odc.service.objectstorage.cloud.CloudResourceConfigurations;
@@ -35,7 +38,7 @@ import com.oceanbase.odc.service.objectstorage.cloud.model.ObjectStorageConfigur
 import com.oceanbase.odc.service.objectstorage.cloud.model.UploadObjectTemporaryCredential;
 
 @Ignore
-public class AlibabaCloudSecurityTokenServiceTest {
+public class AWSSecurityTokenServiceTest {
     private static final String FILE_NAME = "src/test/resources/data/test0001.txt";
     private static final String FILE_NAME_CN_ZH = "src/test/resources/data/中文名称.txt";
     private static ObjectStorageConfiguration configuration;
@@ -43,7 +46,7 @@ public class AlibabaCloudSecurityTokenServiceTest {
 
     @BeforeClass
     public static void beforeClass() {
-        configuration = ITConfigurations.getOssConfiguration();
+        configuration = ITConfigurations.getS3Configuration();
         CloudClient cloudClient = new CloudResourceConfigurations().publicEndpointCloudClient(() -> configuration);
         CloudClient internalCloudClient =
                 new CloudResourceConfigurations().internalEndpointCloudClient(() -> configuration);
@@ -62,17 +65,26 @@ public class AlibabaCloudSecurityTokenServiceTest {
         generateTempCredential("中文名称.txt", FILE_NAME_CN_ZH);
     }
 
-    private void generateTempCredential(String fileName, String filepath) {
+    public void generateTempCredential(String fileName, String filePath) {
         GenerateTempCredentialReq req = new GenerateTempCredentialReq();
         req.setFileName(fileName);
         UploadObjectTemporaryCredential credential = service.generateTempCredential(req);
 
-        OSS oss = new OSSClientBuilder().build(credential.getEndpoint(), credential.getAccessKeyId(),
-                credential.getAccessKeySecret(), credential.getSecurityToken());
-        PutObjectResult putObjectResult =
-                oss.putObject(configuration.getBucketName(), credential.getFilePath(), new File(filepath));
+        AWSSessionCredentials sessionCredentials = new BasicSessionCredentials(
+                credential.getAccessKeyId(), credential.getAccessKeySecret(),
+                credential.getSecurityToken());
 
-        Assert.assertEquals(putObjectResult.getClientCRC(), putObjectResult.getServerCRC());
+        AmazonS3 s3 = AmazonS3ClientBuilder.standard()
+                .withCredentials(new AWSStaticCredentialsProvider(sessionCredentials))
+                .withRegion(credential.getRegion())
+                .disableChunkedEncoding()
+                .build();
+
+        PutObjectResult putObjectResult =
+                s3.putObject(configuration.getBucketName(), credential.getFilePath(), new File(filePath));
+
+        Assert.assertNotNull(putObjectResult.getETag());
     }
+
 
 }
