+++ server/integration-test/src/test/java/com/oceanbase/odc/service/objectstorage/cloud/CloudObjectStorageUtilTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 OceanBase.
+ * Copyright (c) 2024 OceanBase.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.service.cloud;
+package com.oceanbase.odc.service.objectstorage.cloud;
 
 import org.junit.Assert;
 import org.junit.Test;
