+++ server/integration-test/src/test/java/com/oceanbase/odc/service/objectstorage/cloud/S3ObjectStorageServiceIT.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 OceanBase.
+ * Copyright (c) 2024 OceanBase.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,21 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.objectstorage.cloud;
+package com.oceanbase.odc.service.objectstorage.cloud;
 
 import org.junit.Ignore;
 
 import com.oceanbase.odc.ITConfigurations;
-import com.oceanbase.odc.service.objectstorage.cloud.CloudObjectStorageService;
-import com.oceanbase.odc.service.objectstorage.cloud.CloudResourceConfigurations;
 import com.oceanbase.odc.service.objectstorage.cloud.client.CloudClient;
 import com.oceanbase.odc.service.objectstorage.cloud.model.ObjectStorageConfiguration;
 
 @Ignore
-public class OBSObjectStorageServiceIT extends AbstractCloudObjectStorageServiceTest {
+public class S3ObjectStorageServiceIT extends AbstractCloudObjectStorageServiceTest {
     @Override
     CloudObjectStorageService createCloudObjectStorageService() {
-        ObjectStorageConfiguration configuration = ITConfigurations.getOBSConfiguration();
+        ObjectStorageConfiguration configuration = ITConfigurations.getS3Configuration();
         CloudClient cloudClient = new CloudResourceConfigurations().publicEndpointCloudClient(() -> configuration);
         CloudClient internalCloudClient =
                 new CloudResourceConfigurations().internalEndpointCloudClient(() -> configuration);
